# BE-Challenge/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Documentation",
    "version": "1.0.0",
    "description": "API for user authentication, authorization, and management"
  },
  "paths": {
...
    "/create-vietqr": {
      "post": {
        "summary": "Create a VietQR payment QR code",
        "tags": [
          "Payment"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "amount": {
                    "type": "string",
                    "description": "The amount to be paid",
                    "example": "100000"
                  }
                },
                "required": [
                  "amount"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully generated VietQR payment URL",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "qrUrl": {
                      "type": "string",
                      "description": "The URL for the VietQR payment QR code",
                      "example": "https://img.vietqr.io/image/123456-78901234-ABC.png?amount=100000&addInfo=Ma%20giao%20dich%20123456789&accountName=Bank%20Account"
                    },
                    "transactionId": {
                      "type": "string",
                      "description": "The unique transaction ID",
                      "example": "123456789"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body"
          }
        }
      }
    },
    "/check-transaction-status/{transactionId}": {
      "get": {
        "summary": "Check the status of a VietQR payment transaction",
        "tags": [
          "Payment"
        ],
        "parameters": [
          {
            "name": "transactionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The unique transaction ID",
              "example": "123456789"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Transaction status successfully retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "description": "The status of the transaction",
                      "example": "success"
                    },
                    "transaction": {
                      "type": "object",
                      "description": "Details of the transaction",
                      "properties": {
                        "transactionId": {
                          "type": "string",
                          "description": "The transaction ID",
                          "example": "123456789"
                        },
                        "amount": {
                          "type": "string",
                          "description": "The amount of the transaction",
                          "example": "100000"
                        },
                        "description": {
                          "type": "string",
                          "description": "Description of the transaction",
                          "example": "Payment for Ma giao dich 123456789"
                        },
                        "status": {
                          "type": "string",
                          "description": "Current status of the transaction",
                          "example": "completed"
                        }
                      }
                    }
                  },
                  "required": [
                    "status"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Transaction not found"
          },
          "500": {
            "description": "Error checking transaction status"
          }
        }
      }
    },
...
    "/posts": {
      "post": {
        "summary": "Create a new post",
        "tags": [
          "Posts"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  },
                  "userId": {
                    "type": "string"
                  }
                },
                "required": [
                  "title",
                  "content",
                  "userId"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Post created successfully"
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
...
    "/challenge": {
      "get": {
        "summary": "Get all challenges",
        "tags": [
          "Challenge"
        ],
        "responses": {
          "200": {
            "description": "A list of challenges",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "example": "123"
                      },
                      "title": {
                        "type": "string",
                        "example": "Challenge 1"
                      },
                      "description": {
                        "type": "string",
                        "example": "Description of challenge 1"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new challenge",
        "tags": [
          "Challenge"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  }
                },
                "required": [
                  "title",
                  "description"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Challenge created successfully"
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
...
    "/challengeUser": {
      "post": {
        "summary": "Create a new user challenge",
        "tags": [
          "ChallengeUser"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userID": {
                    "type": "string"
                  },
                  "challengeID": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string"
                  }
                },
                "required": [
                  "userID",
                  "challengeID",
                  "status"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User challenge created successfully"
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
...
  }
}